# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "minibank.",
        "title": "minibank",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost",
    "basePath": "/api/v1",
    "paths": {
...
        "/users/{id}/accounts": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Fetch the list of accounts using the provided User ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Retrieve list of accounts based on the provided User ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved account details",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/mapper.Account"
                            }
                        }
                    },
                    "404": {
                        "description": "Accounts not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Create a new account using the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Create a new account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Account details for creation",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/account.AccountCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "A new account has been created with number: {string}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "account.AccountCreateRequest": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string",
                    "example": "810"
                },
                "name": {
                    "type": "string",
                    "example": "Удачный"
                }
            }
        },
...
        "mapper.Account": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string",
                    "example": "40817810902007654321"
                },
                "amount": {
                    "type": "number",
                    "example": 99999.99
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-09-19T10:58:00.000Z"
                },
                "currency": {
                    "type": "string",
                    "example": "810"
                },
                "id": {
                    "type": "string",
                    "example": "fdee7aae-f79f-4653-8a16-9207e6805b93"
                },
                "interest_rate": {
                    "type": "number",
                    "example": 0.125
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "Удачный"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-09-19T10:58:00.000Z"
                },
                "user_id": {
                    "type": "string",
                    "example": "fdee7aae-f79f-4653-8a16-9207e6805b93"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}